Creating Event interfaces
The Event interface must be implemented by every event we want to handle
in the application.

The EventPayload is an empty interface that we can use to represent the
event payload, which may contain information about the entity which this
event is related to.

The EventMetadata is also an empty interface that we can use to
represent any data related to the event itself that defining in the
paylaod may not be the best place.