Creating EventPublisher interface
The EventPublisher interface must be implemented to allow the
application to publish data to a stream such as Apache Kafka or AWS Kinesis.

The EventPublishRequest is an object that groups every information
required to publish the event to the stream. It allows us to set the
stream name, partition key and the data.

The EventException is an abstract class that represents any error that
may happen while dealing with streams. Every stream implementation
should extend it to provide fine grained error handling.