Implementing EventMixIn
The EventMixIn allows us to use Jackson 3 feature named MixIn to work
with polymorphic deserialization. It has the same structure as the Event
interface but contains annotations to instruct ObjectMapper how to
identify what implementation of the Event interface must be used based
on the JSON structure.

The JSONConfiguration is a Spring Context configuration class that
provides an ObjectMapper instance pre-configured to use the EventMixIn.