Enabling stream consumption
Creating Configuration classes to expose schedulers and processors for
each group of events. The configuration is based on values defined in
property file with fallback from the environment.

Each configuration class provides one instance of StreamProperties that
contains the information about the stream that should be consumed.

They also provide one instance of ShardRecordProcessor which handles
records from the stream shard and one instance of Scheduler which
uses the ShardRecordProcessorFactory implementation to find the
appropiate ShardRecordProcessor instance to handle one specific type of
record received from the stream shard.